Fix when one runner

+++ serve.js
@@ -7,13 +7,22 @@ gulp.task("serve:serve", async () => {
         await stopServer();
     }
     server = fork(".");
+    server.on("exit", () => {
+        server = null;
+    });
+    server.on("error", () => {
+        server = null;
+    });
 });
 
 function stopServer() {
     return new Promise((resolve, reject) => {
+        if (server == null) {
+            resolve();
+            return;
+        }
         server.on("exit", resolve);
         server.on("error", reject);
         server.kill("SIGINT");
-        server = null;
     });
 }
